// packages/video/lib/types/ArchiveMode.ts
/**
 * Enum representing different archive modes.
 */
export enum ArchiveMode {
  /**
   * Archive mode set to 'manual', where archives are started and stopped manually.
   */
  MANUAL = 'manual',

  /**
   * Archive mode set to 'always', where archives are continuously recorded.
   */
  ALWAYS = 'always',
}